requirements.txt
openai
langchain
langchain-core
langchain_openai
python-dotenv
langchain-chroma
faiss-cpu
streamlit
streamlit-extras
chromadb
langsmith